refactor: move tip data store to app preference
