+++ app/src/main/java/com/karrar/movieapp/data/repository/TipRepositoryImp.kt
@@ -1,16 +1,16 @@
 package com.karrar.movieapp.data.repository
 
-import com.karrar.movieapp.data.local.TipConfiguration
+import com.karrar.movieapp.data.local.AppPreferences
 import javax.inject.Inject
 
 class TipRepositoryImp @Inject constructor(
-    private val tipConfiguration: TipConfiguration,
+    private val appPreferences: AppPreferences,
 ) : TipRepository, BaseRepository()  {
     override suspend fun getCategoryTipStatus(): Boolean {
-        return tipConfiguration.getCategoryTipStatus()
+        return appPreferences.getCategoryTipStatus()
     }
 
     override suspend fun closeCategoryTip() {
-        tipConfiguration.closeCategoryTip()
+        appPreferences.closeCategoryTip()
     }
 }
