+++ app/src/main/java/com/karrar/movieapp/di/LocalDataSourceModule.kt
@@ -4,8 +4,6 @@ import com.karrar.movieapp.data.local.AppConfiguration
 import com.karrar.movieapp.data.local.AppConfigurator
 import com.karrar.movieapp.data.local.AppPreferences
 import com.karrar.movieapp.data.local.AppPreferencesImpl
-import com.karrar.movieapp.data.local.TipConfiguration
-import com.karrar.movieapp.data.local.TipConfigurator
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -23,9 +21,4 @@ abstract class LocalDataSourceModule {
     @Singleton
     @Binds
     abstract fun bindAppPreference(appPreferencesImpl: AppPreferencesImpl): AppPreferences
-
-
-    @Singleton
-    @Binds
-    abstract fun bindTipConfiguration(tipConfigurator: TipConfigurator) : TipConfiguration
 }
